daily_summaries_csv_gen_job/requirements.txt
pandas
google-cloud-bigtable
google-cloud-storage